Move tolerance and reduction computation out of test
